refactor: split sync/books.rs into subdir (#466)
The file was 945 lines, over the ~800 soft cap from
.claude/rules/05-rust-style.md. Split by sub-topic following the
db/src/books/ precedent: a `mod.rs` orchestrator + per-bucket files
(`new`, `changed`, `removed`, `backfill`) and a `shared.rs` for the
cross-bucket row writers / metadata link dispatch / attach helpers.

Pure mechanical move — no signature, visibility, logic, or `#[cfg]`
changes. `pub use books::{...}` from `sync.rs` is unchanged, so callers
in `server/`, `frontend/`, and sibling `db/` modules see no diff. The
two `pub(super)` sibling-module entry points used by
`sync::audiobooks` (`ghost_book_by_id`, `materialize_new_covers`) are
re-exported from `books/mod.rs` and widened to `pub(in crate::sync)`
inside `shared.rs` so the path `super::books::*` still resolves.

`book_detail.rs` from the issue is already a directory — split happened
in a separate change — so only the sync side is in scope here.

Closes #466
